Refactor IntCode to handle Longs

+++ src/main/scala/adventofcode/y2019/Day2.scala
@@ -20,5 +20,5 @@ object Day2 extends Year2019 {
   printDayPart(2, 100 * noun + verb)
 
   private def runProgram(noun: Int, verb: Int) =
-    new IntCode(intcodeProgram.updated(1, noun).updated(2, verb)).run()
+    new IntCode(intcodeProgram.updated(1, noun.toLong).updated(2, verb.toLong)).run()
 }
